% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/celsius_to_farenheit.R
\name{Celsius_to_Fahrenheit}
\alias{Celsius_to_Fahrenheit}
\title{Celsius to Farenheit conversion test}
\usage{
Celsius_to_Fahrenheit(temp_C)
}
\arguments{
\item{temp_C}{numeric}
}
\value{
numeric
}
\description{
Celsius to Farenheit conversion test
}
\examples{
Celsius_to_Fahrenheit(0) # results in 32F
}
