% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/celsius_to_kelvin.R
\name{Celsius_to_Kelvin}
\alias{Celsius_to_Kelvin}
\title{Celsius to Kelvin conversion}
\usage{
Celsius_to_Kelvin(temp_C)
}
\arguments{
\item{temp_C}{numeric}
}
\value{
numeric
}
\description{
Celsius to Kelvin conversion
}
\examples{
Celsius_to_Kelvin(10) # returns 283.15K
}
