% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ferenheit_to_celsius.R
\name{Fahrenheit_to_Celsius}
\alias{Fahrenheit_to_Celsius}
\title{Fahrenheit to Celsius conversion}
\usage{
Fahrenheit_to_Celsius(temp_F)
}
\arguments{
\item{temp_F}{numeric}
}
\value{
numeric
}
\description{
Fahrenheit to Celsius conversion
}
\examples{
Fahrenheit_to_Celsius(32) #returns 0
}
