% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ferenheit_to_kelvin.R
\name{Fahrenheit_to_Kelvin}
\alias{Fahrenheit_to_Kelvin}
\title{Fahrenheit to Kelvin conversion}
\usage{
Fahrenheit_to_Kelvin(temp_F)
}
\arguments{
\item{temp_F}{numeric}
}
\value{
numeric
}
\description{
Fahrenheit to Kelvin conversion
}
\examples{
Fahrenheit_to_Kelvin(32) #returns 273.15K
}
