% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kelvin_to_farenheit.R
\name{Kelvin_to_Fahrenheit}
\alias{Kelvin_to_Fahrenheit}
\title{Kelvin to Fahrenheit conversion}
\usage{
Kelvin_to_Fahrenheit(temp_K)
}
\arguments{
\item{temp_K}{numeric}
}
\value{
numeric
}
\description{
Kelvin to Fahrenheit conversion
}
\examples{
Kelvin_to_Fahrenheit(273.15) # returns 32F
}
